Rename `framing2.rs` to `framing.rs`
