+++ protocols/v2/codec-sv2/src/lib.rs
@@ -22,15 +22,15 @@ pub use encoder::Encoder;
 pub use encoder::NoiseEncoder;
 
 #[cfg(feature = "noise_sv2")]
-pub use framing_sv2::framing2::HandShakeFrame;
-pub use framing_sv2::framing2::Sv2Frame;
+pub use framing_sv2::framing::HandShakeFrame;
+pub use framing_sv2::framing::Sv2Frame;
 
 #[cfg(feature = "noise_sv2")]
 pub use noise_sv2::{self, Initiator, NoiseCodec, Responder};
 
 pub use buffer_sv2;
 
-pub use framing_sv2::{self, framing2::handshake_message_to_frame as h2f};
+pub use framing_sv2::{self, framing::handshake_message_to_frame as h2f};
 
 #[cfg(feature = "noise_sv2")]
 #[derive(Debug)]
